+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 3.4.0 - 2023-11-09
+
+### Added
+
+* Added support for configuring the parser ([#84](https://github.com/gherking/gherking/issues/39)):
+  * Support for configuring what parametrized tag format should be used during parsing
+
 ## 3.3.2 - 2023-02-15
 
 ### Fixed